Fix database stored procedure call native type mapping

+++ Database/GameServer01/0061_CreateInsertSubpassword.sql
@@ -4,7 +4,7 @@ CREATE PROCEDURE InsertSubpassword(
     IN InPassword VARCHAR(10),
     IN InQuestion INT,
     IN InAnswer VARCHAR(16),
-    OUT OutSuccess TINYINT,
+    OUT OutSuccess INT,
     OUT OutType INT
 )
 BEGIN
